Move the logic for mvoing columns in utils folder
